changed package name, add setuppy, litepolis as core engine for modularization
