+++ litepolis/rbac_policy.csv
@@ -1,0 +1,26 @@
+p, user, /users/profile, GET
+p, user, /users/profile, PUT
+p, user, /users/password, PUT
+p, user, /users/renew, PUT
+
+p, admin, *, *
+
+p, moderator, /users/profile, GET
+p, moderator, /users/profile, PUT
+p, moderator, /users/password, PUT
+p, moderator, /users/renew, PUT
+p, moderator, /comments/*/moderate, GET
+p, moderator, /comments/, PUT
+
+p, *, /login, *
+p, *, /conversations, POST
+p, *, /conversations/*, GET
+p, *, /conversations/*, PUT
+p, *, /comments/, POST
+p, *, /comments/*, GET
+p, *, /comments/*/vote, POST
+
+p, anonymous, /, GET
+
+g, user, users
+g, moderator, users
