--- src/rbac_policy.csv
@@ -1,13 +1,0 @@
-p, user, /dataset1/*, GET
-p, user, /...converation..., GET
-
-p, admin, *, *
-
-p, moderator, /dataset2/resource1, *
-
-p, *, /login, *
-
-p, anonymous, /, GET
-p, anonymous, /...converation..., GET
-
-g, cathy, dataset1_admin
